requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
dimod==0.9.0
dwave-cloud-client==0.7.0
dwave-hybrid==0.4.2
dwave-inspector==0.1.3
dwave-inspectorapp==0.1.3
dwave-neal==0.5.3
dwave-networkx==0.8.4
dwave-ocean-sdk==2.0.1
dwave-qbsolv==0.3.0
dwave-system==0.9.1
dwave-tabu==0.2.2
dwavebinarycsp==0.1.1
Flask==1.1.1
homebase==1.0.1
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.3
jsonschema==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
minorminer==0.1.9
networkx==2.4
numpy==1.18.1
ortools==6.10.6025
penaltymodel==0.16.3
penaltymodel-cache==0.4.1
penaltymodel-lp==0.1.3
penaltymodel-mip==0.2.2
plucky==0.4.3
protobuf==3.6.1
pyparsing==2.3.1
pyqubo==0.4.0
PySocks==1.7.1
python-dateutil==2.8.0
requests==2.23.0
scipy==1.4.1
six==1.12.0
urllib3==1.24.1
Werkzeug==1.0.0
zipp==3.1.0
